tests/wpt/web-platform-tests/css/css-text/parsing/hyphens-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing hyphens with invalid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-hyphens">
<meta name="assert" content="hyphens supports only the grammar 'none | manual | auto'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("hyphens", "normal");
test_invalid_value("hyphens", "none manual");
test_invalid_value("hyphens", "auto auto");
</script>
</body>
</html>
